refactor: Increase uniqueness of virtualized list keys
The current `keyExtractor` appeared to be an odd replacement for
previous usage of Lodash's `identity` utility. It is unclear if this new
`keyExtractor` suffices, or if a more robust key is required. However,
this change was also made in addition the following.

A new `keyExtractor` and `listKey` was required to address the following
error when running tests against the `BlockList`:

```
VirtualizedList trace:
  Child (vertical):
    listKey: 8d343539-3801-4ff1-ac5b-858db2cdf662
    cellKey: 8d343539-3801-4ff1-ac5b-858db2cdf662
  Parent (vertical):
    listKey: rootList
    cellKey: rootList
```